+++ x/evidence/depinject.go
@@ -4,6 +4,7 @@ import (
 	modulev1 "cosmossdk.io/api/cosmos/evidence/module/v1"
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
+	"cosmossdk.io/core/comet"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/depinject/appconfig"
 	eviclient "cosmossdk.io/x/evidence/client"
@@ -30,6 +31,7 @@ type ModuleInputs struct {
 	Environment      appmodule.Environment
 	Cdc              codec.Codec
 	EvidenceHandlers []eviclient.EvidenceHandler `optional:"true"`
+	CometService     comet.Service
 
 	StakingKeeper  types.StakingKeeper
 	SlashingKeeper types.SlashingKeeper
@@ -45,7 +47,7 @@ type ModuleOutputs struct {
 
 func ProvideModule(in ModuleInputs) ModuleOutputs {
 	k := keeper.NewKeeper(in.Cdc, in.Environment, in.StakingKeeper, in.SlashingKeeper, in.AddressCodec)
-	m := NewAppModule(in.Cdc, *k, in.EvidenceHandlers...)
+	m := NewAppModule(in.Cdc, *k, in.CometService, in.EvidenceHandlers...)
 
 	return ModuleOutputs{EvidenceKeeper: *k, Module: m}
 }